requirements.txt
wheel
djangorestframework
django-filter
drf-yasg
djangorestframework-simplejwt
psycopg2-binary
Pillow
python-decouple
python-slugify
celery
redis
peewee
django-cors-headers
django-cleanup
-q -U google-generativeai